♻️ chore(prompter): move steps to its own internal module
