+++ src/cli.py
@@ -6,11 +6,11 @@ from src.installer.installer import Installer
 from src.installer.uv_manager import UvManager
 from src.installer.zsh_configurator import ZshConfigurator
 from src.project_generator.project_generator import ProjectGenerator
-from src.question_prompter.dependencies_step import DependenciesStep
-from src.question_prompter.domain_driven_design_step import DomainDrivenDesignStep
-from src.question_prompter.general_project_step import GeneralProjectStep
 from src.question_prompter.question_wizard import QuestionWizard
-from src.question_prompter.steps import Steps
+from src.question_prompter.step.dependencies_step import DependenciesStep
+from src.question_prompter.step.domain_driven_design_step import DomainDrivenDesignStep
+from src.question_prompter.step.general_project_step import GeneralProjectStep
+from src.question_prompter.step.steps import Steps
 from src.question_prompter.user_requirements import UserRequirements
 
 app = typer.Typer()
