+++ src/question_prompter/step/domain_driven_design_step.py
@@ -1,5 +1,5 @@
 from src.question_prompter.question.free_text_question import FreeTextQuestion
-from src.question_prompter.steps import Step
+from src.question_prompter.step.steps import Step
 from src.question_prompter.template_types import TemplateTypes
